Add database.py
Database script to connect to our postgress database. Define a
sqlalchemy engine and create db session classes to be used on the main
file when creating database instances during CRUD operations. A Base
class is also created that uses the function declarative_base() and
which will be inherited in database models