Update the authentication router
Use fastapi security utility `OAuth2PasswordRequestForm`. As specified in OAuth2 "password flow"
that the client/user must send a username and password fields as form data not JSON.

Update the path operation and use `OAuth2PasswordRequestForm` as a
dependency in the path operation for `/login`

Update the db query since the request form body will contain the
username and password.